Refactor: 마이페이지 코드 분리 (#178)
* Refactor: 마이페이지 코드 분리

* Fix: 마이페이지 플레이리스트 삭제 에러 이슈 수정

* Rename: user.ts로 파일명 변경
