+++ src/hooks/queries/usePlaylistMutation.ts
@@ -1,7 +1,14 @@
-import { useMutation } from "@tanstack/react-query";
+import { useMutation, useQueryClient, InfiniteData } from "@tanstack/react-query";
 import { Video } from "@/types/video";
-import { EditPlaylistPayload, NewPlaylistPayload } from "@/types/playlist";
-import { addVideos, createPlaylist, deleteVideo, editPlaylist } from "@/api/playlist";
+import { EditPlaylistPayload, NewPlaylistPayload, Playlist } from "@/types/playlist";
+import {
+  addVideos,
+  createPlaylist,
+  deleteVideo,
+  editPlaylist,
+  deletePlaylist,
+} from "@/api/playlist";
+import { showToast } from "@/utils/toast";
 
 // 플레이리스트 생성 뮤테이션
 export const useCreatePlaylistMutation = () => {
@@ -30,5 +37,34 @@ export const useDeleteVideoMutation = () => {
   return useMutation({
     mutationFn: ({ playlist_id, video_id }: { playlist_id: string; video_id: string }) =>
       deleteVideo(playlist_id, video_id),
+  });
+};
+
+// 플레이리스트 삭제 뮤테이션
+export const useDeletePlaylistMutation = (userId: string, isOwner: boolean, sortOption: string) => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: deletePlaylist,
+    onSuccess: (deletedId) => {
+      showToast("success", "플레이리스트가 삭제되었습니다.");
+
+      queryClient.setQueryData<InfiniteData<{ playlists: Playlist[]; nextPage: number | null }>>(
+        ["userPlaylists", userId, isOwner, sortOption],
+        (old) =>
+          old
+            ? {
+                ...old,
+                pages: old.pages.map((page) => ({
+                  ...page,
+                  playlists: page.playlists.filter((item) => item.id !== deletedId),
+                })),
+              }
+            : old,
+      );
+    },
+    onError: (error) => {
+      console.error("삭제 실패", error);
+    },
   });
 };
